Make LoaderPage its own file
