+++ src/app/components/form-header/form-header.js
@@ -1,5 +1,7 @@
 import React from 'react';
-import {RawHTML, LoaderPage} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
+
 import './form-header.scss';
 
 function FormHeader({data, prefix}) {
