+++ src/app/components/jsx-helpers/context-loader.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {LoaderPage} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 
 export default function buildContextLoader() {
     const Context = React.createContext();
