+++ src/app/components/shell/footer/footer.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import CookieDialog from '../cookie-notice/cookie-notice';
-import {LoaderPage, RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faFacebookF} from '@fortawesome/free-brands-svg-icons/faFacebookF';
 import {faTwitter} from '@fortawesome/free-brands-svg-icons/faTwitter';
