+++ src/app/pages/campaign/openstax-tutor/openstax-tutor.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import {LoaderPage, RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import './openstax-tutor.scss';
 
 function Blurb({icon, title, description}) {
