+++ src/app/pages/errata-detail/errata-detail.js
@@ -1,4 +1,4 @@
-import {LoaderPage} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import React from 'react';
 import ProgressBar from './progress-bar/progress-bar.jsx';
 import Detail from './detail/detail';
