+++ src/app/pages/errata-summary/hero/hero.js
@@ -1,5 +1,6 @@
 import React, {useState, useEffect} from 'react';
-import {RawHTML, LoaderPage} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faInfoCircle} from '@fortawesome/free-solid-svg-icons/faInfoCircle';
 import bookPromise from '~/models/book-titles';
