+++ src/app/pages/faq/faq.js
@@ -1,4 +1,5 @@
-import {LoaderPage, RawHTML, useToggle} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML, useToggle} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import React from 'react';
 import cn from 'classnames';
 import $ from '~/helpers/$';
