+++ src/app/pages/institutional-partnership-application/institutional-partnership-application.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import {LoaderPage, RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import sample from 'lodash/sample';
 import './institutional-partnership-application.scss';
