+++ src/app/pages/partners/partners.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import {RawHTML, LoaderPage, useToggle} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import {RawHTML, useToggle} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import Controls from './controls/controls';
 import MobileControlRow from './mobile-controls/mobile-controls';
 import Results, {costOptions, equityOptions} from './results/results';
