+++ src/app/pages/research/research.js
@@ -1,4 +1,4 @@
-import {LoaderPage} from '~/components/jsx-helpers/jsx-helpers.jsx';
+import LoaderPage from '~/components/jsx-helpers/loader-page';
 import React from 'react';
 import LazyLoad from 'react-lazyload';
 import HeroSection from './hero-section/hero-section';
